+++ src/components/components.md
@@ -1,0 +1,3 @@
+## Contents
+
+Components that are used and reused across modules should be placed in the components folder.
