+++ src/modules/resources/Resource.tsx
